Improvement of the recursion of the permissions.
The changes that I propose allow within the 'directory' function a more accurate management of the 'user', 'group' and 'mode' options when applied recursively.
When a 'directory' function applies the options marked recursively, they are overwritten by other 'directory' functions if they correspond to a managed subpath. To do this, each directory function auto-requires the 'directory' and 'managed' functions with managed subpath.

In this way no changes will be applied and displayed during the application of a state even if multiple functions interact on the same directory or file. This is not currently the case, if permissions is managed at subpath level, by 'directory' or 'managed' function, the application of the permissions in recursive mode by parents 'directory' function is applied repeatedly with each application of the states.

It is possible to test what has been said with the following states:
# cat test.sls

/tmp/user:
  file.directory:
    - user: root
    - group: root
    - mode: '0755'
    - recurse:
        - user
        - group

/tmp/user/dir:
  file.directory:
    - user: named
    - group: named
    - mode: '0750'
    - recurse:
        - user
        - group
        - mode
        - ignore_dirs

/tmp/user/dir/subdir:
  file.directory

/tmp/user/dir/subfile1:
  file.managed:
    - user: cron
    - group: cron
    - mode: '600'

/tmp/user/dir/subfile2:
  file.managed

/tmp/user/file:
  file.managed:
    - group: mail
    - mode: '0600'